// 2A-finance-tracker/src/app/core/models/transaction.model.ts
//model is a data structure that defines what transaction look like
//only shape of the data no logic


export interface Transaction {
    //interface is a blueprint for an object, transaction is the name of the interface, export makes it usable in other files(services or compoenents)
  id?: number|string;
      //id is optional because it's auto-generated by the database
  property: 'Earthy Escape' | 'Millennium Farm House';
  //we have set property to limit the values to these two strings only
  type: 'income' | 'expense';
  //type is limited to income or expense only
  category: string;
  amount: number;
  date: string;
  description?: string;
}
//example of a transaction object
// {
//   id: 1,
//   property: 'Earthy Escape',
//   type: 'income',
//   category: 'Room Booking',
//   amount: 15000,
//   date: '2026-01-21',
//   description: 'Weekend stay'
// }